# image-vector-service/openapi/image-vector-service.openapi.json
{
  "components": {
    "schemas": {
      "AuditStatusResponse": {
        "properties": {
          "datasetSize": {
            "title": "Datasetsize",
            "type": "integer"
          },
          "indexHash": {
            "title": "Indexhash",
            "type": "string"
          },
          "indexVersion": {
            "title": "Indexversion",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "modelHash": {
            "title": "Modelhash",
            "type": "string"
          },
          "modelVersion": {
            "title": "Modelversion",
            "type": "string"
          },
          "onChainMatch": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onchainmatch"
          },
          "snapshotId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshotid"
          }
        },
        "required": [
          "modelVersion",
          "modelHash",
          "indexVersion",
          "indexHash",
          "datasetSize"
        ],
        "title": "AuditStatusResponse",
        "type": "object"
      },
      "Body_search_similar_images_api_vector_search_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_search_similar_images_api_vector_search_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "SearchResponse": {
        "properties": {
          "indexVersion": {
            "title": "Indexversion",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "modelVersion": {
            "title": "Modelversion",
            "type": "string"
          },
          "queryHash": {
            "title": "Queryhash",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "title": "Results",
            "type": "array"
          },
          "topK": {
            "title": "Topk",
            "type": "integer"
          }
        },
        "required": [
          "method",
          "topK",
          "modelVersion",
          "indexVersion",
          "queryHash",
          "results"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "SearchResult": {
        "properties": {
          "className": {
            "title": "Classname",
            "type": "string"
          },
          "contentHash": {
            "title": "Contenthash",
            "type": "string"
          },
          "distance": {
            "title": "Distance",
            "type": "number"
          },
          "fileName": {
            "title": "Filename",
            "type": "string"
          },
          "gatewayUrl": {
            "title": "Gatewayurl",
            "type": "string"
          },
          "imageId": {
            "title": "Imageid",
            "type": "integer"
          },
          "ipfsUri": {
            "title": "Ipfsuri",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "relativePath": {
            "title": "Relativepath",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "rank",
          "score",
          "distance",
          "imageId",
          "className",
          "fileName",
          "ipfsUri",
          "gatewayUrl",
          "contentHash",
          "relativePath"
        ],
        "title": "SearchResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Image Vector Search Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/vector/audit-metadata": {
      "get": {
        "operationId": "audit_metadata_api_vector_audit_metadata_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Audit Metadata Api Vector Audit Metadata Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Audit Metadata"
      }
    },
    "/api/vector/audit-status": {
      "get": {
        "operationId": "audit_status_api_vector_audit_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "snapshot_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Status"
      }
    },
    "/api/vector/health": {
      "get": {
        "operationId": "health_api_vector_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "boolean"
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "title": "Response Health Api Vector Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/api/vector/rebuild-index": {
      "post": {
        "operationId": "rebuild_index_api_vector_rebuild_index_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Rebuild Index"
      }
    },
    "/api/vector/search": {
      "post": {
        "operationId": "search_similar_images_api_vector_search_post",
        "parameters": [
          {
            "in": "query",
            "name": "top_k",
            "required": false,
            "schema": {
              "default": 10,
              "minimum": 1,
              "title": "Top K",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_search_similar_images_api_vector_search_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Similar Images"
      }
    }
  }
}